Twitter'dan alınan tweetlerin scoreları belirlendi. Backend senaryoları gerçekleştirildi. Query time 15-22 saniye sürmekte optimizasyon gerekiyor

+++ src/main/java/com/burakenesdemir/stockmarket/resource/TweetResource.java
@@ -22,5 +22,5 @@ public class TweetResource implements Serializable {
     private String time;
 
     @JsonProperty("sentiment")
-    private Integer sentiment;
+    private Float sentiment;
 }
